Set proper system openssldir in package build

+++ Builds/containers/shared/build_deps.sh
@@ -32,7 +32,8 @@ wget https://www.openssl.org/source/openssl-${OPENSSL_VER}.tar.gz
 tar xf openssl-${OPENSSL_VER}.tar.gz
 cd openssl-${OPENSSL_VER}
 # NOTE: add -g to the end of the following line if we want debug symbols for openssl
-./config -fPIC --prefix=/opt/local/openssl --openssldir=/opt/local/openssl zlib shared
+SSLDIR=$(openssl version -d | cut -d: -f2 | tr -d [:space:]\")
+./config -fPIC --prefix=/opt/local/openssl --openssldir=${SSLDIR} zlib shared
 make -j$(nproc)
 make install
 cd ..
@@ -40,36 +41,38 @@ rm -f openssl-${OPENSSL_VER}.tar.gz
 rm -rf openssl-${OPENSSL_VER}
 LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/opt/local/openssl/lib /opt/local/openssl/bin/openssl version -a
 
-cd /tmp
-wget https://github.com/doxygen/doxygen/archive/Release_1_8_14.tar.gz
-tar xf Release_1_8_14.tar.gz
-cd doxygen-Release_1_8_14
-mkdir build
-cd build
-cmake -G "Unix Makefiles" ..
-make -j$(nproc)
-make install
-cd ../..
-rm -f Release_1_8_14.tar.gz
-rm -rf doxygen-Release_1_8_14
+if [ "${CI_USE}" = true ] ; then
+    cd /tmp
+    wget https://github.com/doxygen/doxygen/archive/Release_1_8_14.tar.gz
+    tar xf Release_1_8_14.tar.gz
+    cd doxygen-Release_1_8_14
+    mkdir build
+    cd build
+    cmake -G "Unix Makefiles" ..
+    make -j$(nproc)
+    make install
+    cd ../..
+    rm -f Release_1_8_14.tar.gz
+    rm -rf doxygen-Release_1_8_14
 
-mkdir -p /opt/plantuml
-wget -O /opt/plantuml/plantuml.jar https://downloads.sourceforge.net/project/plantuml/plantuml.jar
+    mkdir -p /opt/plantuml
+    wget -O /opt/plantuml/plantuml.jar https://downloads.sourceforge.net/project/plantuml/plantuml.jar
 
-cd /tmp
-wget https://github.com/linux-test-project/lcov/releases/download/v1.13/lcov-1.13.tar.gz
-tar xfz lcov-1.13.tar.gz
-cd lcov-1.13
-make install PREFIX=/usr/local
-cd ..
-rm -r lcov-1.13 lcov-1.13.tar.gz
+    cd /tmp
+    wget https://github.com/linux-test-project/lcov/releases/download/v1.13/lcov-1.13.tar.gz
+    tar xfz lcov-1.13.tar.gz
+    cd lcov-1.13
+    make install PREFIX=/usr/local
+    cd ..
+    rm -r lcov-1.13 lcov-1.13.tar.gz
 
-pip install requests
-pip install https://github.com/codecov/codecov-python/archive/master.zip
+    pip install requests
+    pip install https://github.com/codecov/codecov-python/archive/master.zip
 
-set +e
-mkdir -p /opt/local/nih_cache
-mkdir -p /opt/jenkins
-set -e
+    set +e
+    mkdir -p /opt/local/nih_cache
+    mkdir -p /opt/jenkins
+    set -e
+fi
